requirements.txt
flask==3.0.0
stripe==7.8.0
requests==2.31.0
ecdsa==0.18.0
python-dotenv==1.0.0
cryptography==41.0.7
schedule==1.2.0
flask-babel==4.0.0
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.7
beautifulsoup4==4.12.2
selenium==4.15.0
pyotp==2.8.0
qrcode==7.4.2
pillow==10.0.0
pytest==7.4.0
flake8==6.0.0
smtplib
email
